Transfrom pids_arr to actual pids before calling sys_restart
Allow 'mktree' to restart a tree of processes without the need to
clone-with-pid by modifying the pids_arr structure prior to the
actual restart.

We use a "helper task" that forks the init (that forks the others),
and expects to receive via a pipe the translations old ==> new pids.
The task also reads the checkpoint in and write the (modified) data
out to the stdin of the restarting process.

Of course, this assumes that the application can tolerate the change
of pids after a restart.