no-pids: use child as feeder to avoid orphaning the group
In 'no-pids' mode, use the child process for the feeder, so that
the original task (mktree) remains the process group leader and
isn't orphaned.

The feeder task is made reparented to init(1) so the original
task need not collect it (if restart succeeds, of course it won't !)

Add documentation and do some code cleanups.