Inroduce DumpForest algorithm to build creator-child-relationship tree
Support restart of multiple processes from userspace, in the presence
of threads, dead (zombie) tasks and clones with CLONE_PARENT, and while
preserving the session-id.

The algorithm was described in the paper "Transparent Checkpoint/Restart
of Multiple Processes on Commodity Operating Systems" in USENIX 2007
(http://www.usenix.org/events/usenix07/tech/full_papers/laadan/laadan_html/paper.html).

We distinguish between a process parent and its creator: the differ for
threads and for clone with CLONE_PARENT. For each task the algorithm
assigns a creator, and for each creator the algorihtm assigns the list
of children to create. In the next patch this will be used as a set of
instructions to create the processes

The output is a tree that describes these relationships. The tasks are
stores in a large array and the tree structure is imposed on this array.
The size array is three times the number of tasks, to accommodate for
new place-holder tasks that the algorithm may add on demand.

To provide fast lookup of a task in the array/tree given its pid, there
is also a simple hash table implementation that translates a pid to the
pointer to the task.

If restarting a non-container (that is, a subtree of processes from some
namespace), then the "root" process is not init. The code is careful to
compare with the subtree root instead. On the other hand, we can't have
zombie processes.