mktree: set 'session->phantom' when adding a placeholder task
Placeholder tasks are used to generate session ids of tasks that
have died prior to the checkpoint. There is at most one placeholder
per session, pointed to by ->phantom.

Also, Sukadev Bhattiprolu observed that a task with ppid==1 and a
session id like the container-init can be viewed either as an orphan
task or as a child of containter-init. To avoid an unnecessary
placeholder in this case, switch the order of the orphan and the
same-sid-as-parent tests.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>