mktree: use a coordinator task during restart
Kernel code was updated to expect a coordinator task when multiple
tasks are restarted (commit b0ddfc4784a543fdb98662d975963468e13ad3fa)

This patch adapts mktree.c to that scheme. In particular, the main
task is the coordinator. It now forks the feeder task and then also
forks the root restarting task.

One benfit is that error reporting makes sense now: the user will
see a suitable error message instead of the segfault of the init
task.

Also, it's now easier to add the "CLONE_NEWPID" when creating the
init task to put the restarting processes in a new namespace.

Added switch to 'mktree':  --wait (-w) to wait for restart (root)
task to exit, --show-status to show its exit status, --copy-status
to imitate its exit status, and --no-wait (-W) to not wait for it.

Catch SIGINT and pass it to the restart task if in --wait mode.

Catch SIGCHLD (until restart succeeds) to detect errors.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>