mktree.c: now works with --pidns (new pid namespace)
In particular, for --pidns mktree checks the pid of the root_task.

If the pid is 1, proceed as before for a new namespace, and fork
the root_task with CLONE_NEWPID.

If the pid isn't 1, then fork a coordinator task into a new pidns
so that it becomes the init task there, and don't use CLONE_NEWPID
for the root task.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>