mktree.c: ensure that all tasks are ready for coordinator
The (recent) restart logic requires that all restarting tasks be
already alive when the coordinator starts to read the checkpoint image
file (i.e. prior to calling prepare_descendants() in the kernel).

This has already been the case for the "no --pids" case, because the
feeder would first wait to hear from all restarting tasks about their
new pids, and only then push actual data to the kernel.

Now, for the '--pids' case, piggy-back on this mechanism to create the
same end effect, so when coordinator starts to process data from the
file descriptor, the hierarchy is properly prepared.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>