mktree.c: make feeder task a thread instead of child
When mktree is also the coordinator task and also a container init,
e.g. when started with 'ns' and with '--no-pidns', the feeder process
would be re-parented to it and produce a SIGCHLD when it terminated.
In turn, this would interrupt coordinator's wait-for-completion of
the restart.

Converting the feeder task to a thread (but without using SIGCHLD in
the flags) eliminates this problem.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>