Use __WALL option in waitpid()
The coordinator creates the children with clone(), so pass the __WALL
option to waitpid(). Otherwise the waitpid() will fail, and bring the
container down even as the application is restarting successfully :-)

For instance, if the original app is executed as:
  $ ns_exec -cpuimP pid.ptree1 -- ./ptree1 -n 2 -d 1 -w sleep

Then ns_exec creates the app in new container, with p->exit_signal 0
for the container-init. Restart correctly restores the exit_signal to
0. Because of that, we need the coordinator to use __WALL.

Signed-off-by: Sukadev Bhattiprolu <sukadev@us.ibm.com>
[orenl@cs.columbia.edu: use __WALL in all waitpid() of coordinator]