mktree: cleanup collect of coordinator/root task
Fix the logic that collects the coordinator and/or root task
and report their exit status.

This mainly fixes bad behavior when 'mktree' itself is a container
init (e.g. started with nsexec), in which case it collects not only
the coordinator but all orphan zombies resulting from the restart.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>