mktree: add option '--freezer=' to freeze tasks after restart
Use this option to freeze the resulting task hierarchy once the
restart succeeds:
	$ mktree --freezer=/cgroup/name < ckpt.image

With this option, mktree expects a freezer cgroup at the path given as
an argument. It will add all restarting tasks to that cgroup, and will
use the RESTART_FROZEN restart flag to request that this cgroup be
frozen before processes are allowed to resume execution.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>