restart: fix restart of with multithreaded root_task
This patch fixes a restart failure when the root task has multiple
threads. Serge Hallyn reported:

  "hmm, on process-tree/run-pthread.sh, I now get 'pid 2: invalid ppid
  0' on restart .... it uses pthread_clone(), which results in a bunch
  of threads with pid 1 and ppid 0. restart hates that, bc the clones
  are not container init"

The fix is to handle threads before the test for a ppid==0, because
threads are always created by the thread group leader.

Also add a test when handling threads - and fail if the thread group
leader is missing (previously it would ... crash).

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>