restart: correct handling of root task with sid != pid
The case where root_sid != root_pid is handles explicitly. It can only
happen when the checkpoint was of a subtree (in container, root_sid is
either root_pid or 0).  This means that root_sid was inherited from an
ancestor of that subtree. So we make root_task here also inherit its
sid from its ancestor (whatever the 'restart' process had).

We do it by forcing it to be 0. We also need to force all references
to it from other processes, via sid and pgid, to 0. For that, we keep
the root_sid to compare against (see one-line comment below).

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>