restart: respond to (mimic) signal (if --copy-status asked)
In --copy-status mode, the coordinator waits for the restarted
root-task and mimics the way it exits: either by exiting with the same
exit status, or by sending the same fatal signal to self.

In the latter case of the signal, we need to ensure that the default
behavior of that signal is set, or may not kill us.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>