restart: fix sqsuash of sid/pgid to coordinator's sid
In subtree restart some sid/pgid may be zeroes to indicate that
they should become the coodinator's sid. These tasks are marked
in the tasks_array with flags.

The actual coversion was incorrect and causes a buffer overrun,
which may end up in the data not being set up correctly for the
kernel.

This patch fixes this by pulling the substitution to a separate
loop.

This fixes a restart failure reported by Serge Hallyn where the
kernel was seeing a pid that was not swapped (by the feeder).

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>