user-cr: add nsexeccwp to test eclone
One of the concerns with eclone is whether the stack handling is all
correct and robust enough to withstand real usage.  Little testcases
playing with pid values are also necessary, but can't replace really
using clone-with-pids to start a shell from which to keep working.

This patch tweaks the old ns_exec.c namespace manipulation program to
add a -z option to specify a pid.  So you can:

    nsexeccwp -cmp /bin/bash # start a shell in a new pidns+mntns
    mount -t proc proc /proc # mount private /proc
    echo $$
                1
    nsexeccwp -z /bin/bash   #  start a shell with pid 999
    echo $$
            999

[ ntl - minor updates to original version to use clone_args/eclone() ]

Signed-off-by: Serge E. Hallyn <serue@us.ibm.com>
Signed-off-by: Nathan Lynch <ntl@pobox.com>