restart: close open files when coordinator is container init (--pidns)
A restart into a new pidns will make the coordinator the init(1) of
the new pidns if the pid of the (restart) root task in the image is
not 1.

In this case, the coordinator should close all file decsriptors that
link it to the outside world once the restart succeeds. Otherwise,
a subsuquent full (non-subtree) checkpoint will fail due to resource
leaks outside the container.

This behavior is disabled if debugging it turned on to not lose the
output coming from that coordinator.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>