restart: fix --wait for coordinator that becomes container init
Serge Hallyn reported that a "restart --wait --pidns" will not
actually wait in the case of the coordinator becoming the init(1) of
the new pidns. In turn this happens when restarting a
subtree-checkpoint (pid of root task is not 1).

This patch fixes this problem by calling to ckpt_collect_child()
afte the coordinator reports "succeess", if the --wait option was
requested.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>