user-cr: invoke exit system call directly from ckpt_do_feeder
The feeder thread can cause the restart process to fail by indirectly
calling exit_group, which sends SIGKILL to all other threads in the
process.  If the feeder thread "wins" the race, the restart is
disrupted.  A common symptom of this race is the coordinator task
returning from the wait_for_completion_interruptible in
wait_all_tasks_finish with a signal (the SIGKILL) pending.

Calling _exit isn't enough; see
http://www.kernel.org/doc/man-pages/online/pages/man2/exit.2.html#NOTES

Exit the feeder thread by using the syscall() macro.

Signed-off-by: Nathan Lynch <ntl@pobox.com>