user-cr: add generic stack API
A few problems exist with the management of stacks for clone/eclone in
the restart utility.

* Stacks are allocated from the heap via malloc.  This means any
  mishandling of the the stack setup by the platform eclone code can
  result in silent data corruption.  At a time when architecture
  support for eclone is an ongoing effort, this seems unwise...

* Determining a suitable stack pointer to pass to clone(2) is
  platform-dependent; we don't handle e.g. parisc where the stack
  grows upwards.

* There are at least two cases (ckpt_coordinator_pidns and
  ckpt_fork_feeder) where the stack pointer provided to
  clone(2) is outside of the region allocated.

To address these issues, introduce an opaque "genstack" (generic
stack) type and accessor functions.  Stacks are allocated via mmap()
and guard pages are instantiated at both ends.  The accessor functions
provide suitable values for passing to clone(2) and eclone(2) (via
struct clone_args in the latter case).

Signed-off-by: Nathan Lynch <ntl@pobox.com>