user-cr: make powerpc eclone stack setup more robust
For ppc64, it's not sufficient to merely align the stack pointer
(clone_args->child_stack); we must also set up a stack frame (48
bytes).  Encapsulate the stack setup details in a helper function
(stack_setup_frame) which makes appropriate adjustments.

Although the genstack implementation provides zeroed stack pages, it's
best not to rely on that implementation detail, so explicitly zero the
back chain pointer in stack_setup_frame as well.

Signed-off-by: Nathan Lynch <ntl@pobox.com>