restart: remount /proc for new tasks created with CLONE_NEWPID
Not doing this can be a pain for restarted software which relies
on /proc...

This builds on a patch by Serge Hallyn, but also aims to address the
future cases of hierarchical pid-ns:

1) Before mounting the new /proc, first umount the old one, which
isn't necessary anymore. If not already mounted, issue a warning by
default (user can select to skip, warn, or fail).

2) Perform the unshare() together with the remount of /proc, so it
will occur for every new pid-ns and not only for the first one.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>