restart: Remove args->logfile
The args->logfile field is used to store the the name of the logfile to
which the kernel must write its output. If we move the code that opens this
args->logfile from main() to parse_args() and use the existing args->klogfd
field, we can remove the ->logfile field from struct args.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>