restart: Remove args->input
Having both ->input and ->infd in 'struct args' is redundant.

The args->input field is used to store the name of the input file
(checkpoint-image).  If we move the code that opens this file from
main() to parse_args() and use the existing args->infd field, we
can remove the ->input field from struct args.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>