Do not overwrite caller's stdin file descriptor
As indirectly noted by Suka, we used to dup2 the input fd onto
the STDIN_FILENO, and this overwrites the caller's original fd.
This is an undesired outcome from a library...

This patch changes the behavior:
1) The coordinator uses args->infd instead of dup2 and stdin.
2) The feeder still uses stdin/stdout, but that's ok because
 it is cloned with a separate file descriptor table
3) Added sanity checks to assert that only the feeder indeed
 calls those functions that use stdin/stdout

This is the first patch in a series to ensure that the library
does not impact the caller (fds, memory, chdir, etc).

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>