Do not exit() blindly on errors, return error to caller instead
We converted the restart code to become a library and API for
user-cr. When it encounteres an error, it should not just exit
unconditionally; instead it should return an error value to the
caller, and cleanup any mess it may have done.

Children processes under the coordinator, however, may freely
call exit because they do not affect the caller.

The feeder is careful to call sys_exit and not sys_group_exit.

If an error occurs, then cr_restart() will perform some cleanup
to close file descriptors that it opened, and to free memory
that it allocated.

This is the third patch in a series to ensure that the library
does not impact the caller.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>