Remove call to setpgrp()
The purpose of the call was that ckpt_abort() could kill all the
tasks participating in the restart at once through their pgid.

But in doing so it affects the caller (to the library) by modifying
the current pgid ... which we don't want to do.

Besides, some of the restarting tasks call setsid() which will also
change their pgrp, so this method of cleanup was flawed in the first
place.

Finally, all restarting tasks already use prctl() with PR_DEATH_SIGNAL
and probing for their real parent to detect when the parent is gone
and respond by exiting. So it's enough for ckpt_abort() to only kill
the root task - the rest will follow...

This is the forth patch in a series to ensure that the library does
not impact the caller.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>