Initialize of args->{outfd,logfd,infd} in main c/r programs
The initialization belongs to the caller of the cr-library - i.e.
restart-main and checkpoint-main, and not in the cr-library.

Also add a flag '-f' to force overwrite of output files (e.g.
checkpoint image or log output) if they already exists.

Also fix the exit/return value of restart-main.c on failure.

Signed-off-by: Oren Laadan <orenl@cs.columbia.edu>