Avoid redundant definitions
Use a dictionary to collapse multiple definitions of the same enum --
we don't want to output it twice because that will cause a compilation
error. This can happen if we have a comment or other poorly-formatted code
which matches one of the regexes.

Later it might be good to track the values and ensure they match before
collapsing the definitions. If they don't match *then* we should trigger
a compilation error.

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>